Separate out path constants and link create campaign button
