+++ frontend/src/components/RouteConstants.jsx
@@ -1,0 +1,6 @@
+export const rootPath = "/";
+export const userProfilePath = "/profile";
+export const createCampaignPath = "/create-campaign";
+
+export const campaignDetailPath = "campaign/:id";
+export const buildCampaignPath = (campaignId) => `campaign/${campaignId}`;
